huge commit, preparing the base for 64bits support

+++ src/emu/constants.rs
@@ -1,18 +1,18 @@
 
 // https://docs.microsoft.com/en-us/openspecs/windows_protocols/ms-erref/596a1078-e883-4972-9bbc-49e60bebca55
 #![allow(dead_code)]
-pub const STATUS_SUCCESS:u32 = 0x00000000;
-pub const STATUS_ACCESS_DENIED:u32 = 0xC0000022;
-pub const STATUS_INVALID_HANDLE:u32 = 0xC0000008;
-pub const STATUS_NO_MEMORY:u32 = 0xC0000017;
-pub const STATUS_ACCESS_VIOLATION:u32 = 0xC0000005;
-pub const STATUS_INVALID_PARAMETER:u32 = 0xC000000D;
+pub const STATUS_SUCCESS:u64 = 0x00000000;
+pub const STATUS_ACCESS_DENIED:u64 = 0xC0000022;
+pub const STATUS_INVALID_HANDLE:u64 = 0xC0000008;
+pub const STATUS_NO_MEMORY:u64 = 0xC0000017;
+pub const STATUS_ACCESS_VIOLATION:u64 = 0xC0000005;
+pub const STATUS_INVALID_PARAMETER:u64 = 0xC000000D;
 
 pub const NUM_BYTES_TRACE:usize = 16;
-pub const VERSION:u32 = 0x1db10106;
+pub const VERSION:u64 = 0x1db10106;
 
-pub const WAIT_TIMEOUT:u32 = 0x00000102;
-pub const WAIT_FAILED:u32 = 0xFFFFFFFF;
+pub const WAIT_TIMEOUT:u64 = 0x00000102;
+pub const WAIT_FAILED:u64 = 0xFFFFFFFF;
 
 //vectored exception handler
 pub const CALL_FIRST:u32 = 1;
@@ -44,11 +44,11 @@ pub const INTERNET_OPTION_DATA_RECEIVE_TIMEOUT:u32 = 8;
 pub const INTERNET_OPTION_DATA_SEND_TIMEOUT:u32 = 7;
 
 // https://docs.microsoft.com/en-us/windows/win32/wininet/api-flags
-pub const INTERNET_FLAG_SECURE:u32 = 0x00800000;
+pub const INTERNET_FLAG_SECURE:u64 = 0x00800000;
 
-pub const ERROR_NO_MORE_FILES:u32 = 18;
-pub const CREATE_SUSPENDED:u32 = 0x00000004;
-pub const EXCEPTION_EXECUTE_HANDLER:u32 = 1;
+pub const ERROR_NO_MORE_FILES:u64 = 18;
+pub const CREATE_SUSPENDED:u64 = 0x00000004;
+pub const EXCEPTION_EXECUTE_HANDLER:u64 = 1;
 
 pub const PAGE_NOACCESS:u32 = 0x01;
 pub const PAGE_EXECUTE:u32 = 0x00;
@@ -66,41 +66,41 @@ pub const MEM_PRIVATE:u32 = 0x20000;
 
 //// LINUX ////
 
-pub const ENOTSOCK:u32  =       -1i32 as u32;      /* not open sock */
-pub const EPERM:u32     =       -1i32 as u32;      /* permissions error */
-pub const ENOENT:u32    =       -2i32 as u32;      /* No such file or directory */
-pub const ESRCH:u32     =       -3i32 as u32;      /* No such process */
-pub const EINTR:u32     =       -4i32 as u32;      /* Interrupted system call */
-pub const EIO:u32       =       -5i32 as u32;      /* I/O error */
-pub const ENXIO:u32     =       -6i32 as u32;      /* No such device or address */
-pub const E2BIG:u32     =       -7i32 as u32;      /* Argument list too long */
-pub const ENOEXEC:u32   =       -8i32 as u32;      /* Exec format error */
-pub const EBADF:u32     =       -9i32 as u32;      /* Bad file number */
-pub const ECHILD:u32    =      -10i32 as u32;      /* No child processes */
-pub const EAGAIN:u32    =      -11i32 as u32;      /* Try again */
-pub const ENOMEM:u32    =      -12i32 as u32;      /* Out of memory */
-pub const EACCES:u32    =      -13i32 as u32;      /* Permission denied */
-pub const EFAULT:u32    =      -14i32 as u32;      /* Bad address */
-pub const ENOTBLK:u32   =      -15i32 as u32;      /* Block device required */
-pub const EBUSY:u32     =      -16i32 as u32;      /* Device or resource busy */
-pub const EEXIST:u32    =      -17i32 as u32;      /* File exists */
-pub const EXDEV:u32     =      -18i32 as u32;      /* Cross-device link */
-pub const ENODEV:u32    =      -19i32 as u32;      /* No such device */
-pub const ENOTDIR:u32   =      -20i32 as u32;      /* Not a directory */
-pub const EISDIR:u32    =      -21i32 as u32;      /* Is a directory */
-pub const EINVAL:u32    =      -22i32 as u32;      /* Invalid argument */
-pub const ENFILE:u32    =      -23i32 as u32;      /* File table overflow */
-pub const EMFILE:u32    =      -24i32 as u32;      /* Too many open files */
-pub const ENOTTY:u32    =      -25i32 as u32;      /* Not a typewriter */
-pub const ETXTBSY:u32   =      -26i32 as u32;      /* Text file busy */
-pub const EFBIG:u32     =      -27i32 as u32;      /* File too large */
-pub const ENOSPC:u32    =      -28i32 as u32;      /* No space left on device */
-pub const ESPIPE:u32    =      -29i32 as u32;      /* Illegal seek */
-pub const EROFS:u32     =      -30i32 as u32;      /* Read-only file system */
-pub const EMLINK:u32    =      -31i32 as u32;      /* Too many links */
-pub const EPIPE:u32     =      -32i32 as u32;      /* Broken pipe */
-pub const EDOM:u32      =      -33i32 as u32;      /* Math argument out of domain of func */
-pub const ERANGE:u32    =      -34i32 as u32;      /* Math result not representable */
+pub const ENOTSOCK:u64  =       -1i64 as u64;      /* not open sock */
+pub const EPERM:u64     =       -1i64 as u64;      /* permissions error */
+pub const ENOENT:u64    =       -2i64 as u64;      /* No such file or directory */
+pub const ESRCH:u64     =       -3i64 as u64;      /* No such process */
+pub const EINTR:u64     =       -4i64 as u64;      /* Interrupted system call */
+pub const EIO:u64       =       -5i64 as u64;      /* I/O error */
+pub const ENXIO:u64     =       -6i64 as u64;      /* No such device or address */
+pub const E2BIG:u64     =       -7i64 as u64;      /* Argument list too long */
+pub const ENOEXEC:u64   =       -8i64 as u64;      /* Exec format error */
+pub const EBADF:u64     =       -9i64 as u64;      /* Bad file number */
+pub const ECHILD:u64    =      -10i64 as u64;      /* No child processes */
+pub const EAGAIN:u64    =      -11i64 as u64;      /* Try again */
+pub const ENOMEM:u64    =      -12i64 as u64;      /* Out of memory */
+pub const EACCES:u64    =      -13i64 as u64;      /* Permission denied */
+pub const EFAULT:u64    =      -14i64 as u64;      /* Bad address */
+pub const ENOTBLK:u64   =      -15i64 as u64;      /* Block device required */
+pub const EBUSY:u64     =      -16i64 as u64;      /* Device or resource busy */
+pub const EEXIST:u64    =      -17i64 as u64;      /* File exists */
+pub const EXDEV:u64     =      -18i64 as u64;      /* Cross-device link */
+pub const ENODEV:u64    =      -19i64 as u64;      /* No such device */
+pub const ENOTDIR:u64   =      -20i64 as u64;      /* Not a directory */
+pub const EISDIR:u64    =      -21i64 as u64;      /* Is a directory */
+pub const EINVAL:u64    =      -22i64 as u64;      /* Invalid argument */
+pub const ENFILE:u64    =      -23i64 as u64;      /* File table overflow */
+pub const EMFILE:u64    =      -24i64 as u64;      /* Too many open files */
+pub const ENOTTY:u64    =      -25i64 as u64;      /* Not a typewriter */
+pub const ETXTBSY:u64   =      -26i64 as u64;      /* Text file busy */
+pub const EFBIG:u64     =      -27i64 as u64;      /* File too large */
+pub const ENOSPC:u64    =      -28i64 as u64;      /* No space left on device */
+pub const ESPIPE:u64    =      -29i64 as u64;      /* Illegal seek */
+pub const EROFS:u64     =      -30i64 as u64;      /* Read-only file system */
+pub const EMLINK:u64    =      -31i64 as u64;      /* Too many links */
+pub const EPIPE:u64     =      -32i64 as u64;      /* Broken pipe */
+pub const EDOM:u64      =      -33i64 as u64;      /* Math argument out of domain of func */
+pub const ERANGE:u64    =      -34i64 as u64;      /* Math result not representable */
